+++ packages/twenty-server/src/modules/workflow/workflow-executor/workflow-actions/record-crud/utils/parse-finite-number-input.util.ts
@@ -1,0 +1,17 @@
+import { isDefined } from 'twenty-shared/utils';
+
+export const parseFiniteNumberInput = (
+  value: number | string | undefined,
+): number | undefined => {
+  if (!isDefined(value)) {
+    return undefined;
+  }
+
+  if (typeof value === 'string' && value.trim() === '') {
+    return undefined;
+  }
+
+  const parsedValue = Number(value);
+
+  return Number.isFinite(parsedValue) ? parsedValue : undefined;
+};
